Added default value mechanism for variable parameters
It is now possible to define a default value for each variable parameter function upon registering the variable.
Parameter functions will now be called for each evaluation, even if no value was passed.
In this case, the default value will now be passed as the parameter function argument.
Changed index.php and test.mdd accordingly.

Also, changed VariableConverter to not use Closure::fromCallable because of its PHP 7.1 requirement.